overhaul of data truncation
when earliest common year is used in the timespan (default), and we do not use predictors from previous year, treeclim will now use the full data set by padding the climate data with one year of NAs to allow for creation of the parameter matrix. furthermore, warnings have been changed to simple messages.